requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
fastapi==0.115.6
functions==0.7.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
joblib==1.4.2
jsonify==0.5
nltk==3.9.1
numpy==2.2.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
schemas==0.7.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
statsmodels==0.14.4
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
Flask==3.0.0
Flask-Cors==4.0.0
requests==2.31.0
Flask-Login==0.6.3
gunicorn==21.2.0
